Ugly workaround for sprite drawing issues.
sf::Sprite does the most stupid thing imaginable: It applies a fixed (+0.375, +0.375) translation to what it's rendering to achieve *pixel perfect* rendering. With the ugly side effect that for scaled sprites this translation can easily grow above pixel-level where the effect is visible as a sprite that is moved an apparently arbitrary amount. The workaround gets rid of this stupid glTranslatef(...) call.