Adding draw methods where applicable
The SFML2 API adds a virtual function draw that must be created
for classes that inherit from them. I haven't really taken advantage of
anything that this functionality has to offer, but in the interest
of updating our usage of the API, I have put dummy functions where
necessary that call the existing "Render" function in those classes.
I guessed that this "Render" function probably performed the same
job as a "draw" function, and it takes mostly the same parameters.
The addition of a DrawStates parameter to the draw function goes
unused currently.