Changing some logic for the Sky Sprite
I'm marking this out as a separate commit because I'm not sure
whether it's correct. Specifically, the addition of the SetImage
function seems like a bit of a hack because the images could all
be stored as textures in the BitmapManager. Also, the changing of
the Resize function to setScale seems as though it will most
likely produce incorrect results. Nevertheless I'm going to
try to get this compiling and then return to it later in case
it doesn't work.