Fixing Item/Floor for SFML2
I'm much less sure that this is the correct change, so I've left
the original code commented in case they are useful in the future,
but I should clean those up eventually.

This code pattern happens in other files as well, so I'll need to
do some testing to see if it's correct.