Fix GUI texturing and time window layout
Enable texturing before rendering the UI and reset the OpenGL states
before attempting to render the UI. Unfortunately, libRocket is pretty
buggy and unreliable. Adjust the RML file for the time window a bit such
that the message, funds, and population text is displayed again.
Apparently "overflow: hidden" means "don't bother drawing anything at
all". It's time we switched to our own GUI implementation. There's not
a lot of that needed in SimTower anyway.