Making game information more readable
When reading the code for dealing with events I was very confused
by what the numbers refered to, so I made an enum to give meaningful
names to the numbers.
I put it in the OT namespace but not the Item namespace because I
didn't want to prefix that in front of all of the names in Game.cpp.