Fix elevator rendering
The position of the Item is in game coordinates, and the
position of the sprite is in pixel coordinates.
There were some more conflicts with regards to this. I also
fixed some things to make the elevator rendering line up properly.

I'm fairly sure that this code can only handle elevators that are
started at the lobby. This will need to be fixed at some point.