Fixing math in getMouseRegion for Items
The game was having trouble figuring out which item
was underneath the cursor because this function was
returning the wrong type of rectangle. This also makes
it clear that the position is in game coordinates and
not pixel coordinates.
This fixes the extending of elevators, as well as some
other things I'm sure.