Fix positioning of elevators
Somewhere along the line (I think just with the SFML interface changes)
elevators stopped moving, and this unifies the code that moves the
elevator sprite and uses only the altitude variable to determine
where that sprite should go.
This also works for elevators that have cover basement floors, which
is nice.