Fix conditional rendering of items
This section of code has absolutely baffled me for days. In the end,
it was a simple change of coordinate system that was throwing me off.
I have added a large comment explaining why the confusing change that
I've added was necessary. Hopefully this discovery about inconsistent
coordinate systems can be reconciled in the future. I suspect that
this may have changed from SFML 1.6 to SFML 2.0+