Implements Condo occupant movements
There is a lot of this code that is kind of copy pasted
from the Office setup. I could probably share some of this code,
especially if I wanted to jitter arrival times with the office
people as well, and if there end up being other items that we
wish to have similar setups for.